- Add Serverless - Use prebuilt ffmpeg - Use Docker to run build/deploy (as an attempt to make node-canvas work by building on amazonlinux) => Not yet working
