+++ .awscredentials.sample
@@ -1,0 +1,5 @@
+# Fill in your credentials here, and then copy this file to .awscredentials
+[p5-replay]
+region = us-east-1
+aws_access_key_id = YOUR_ACCESS_KEY_ID
+aws_secret_access_key = YOUR_SECRET_ACCESS_KEY
