+++ docker_support_build_and_deploy.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+# Only to be run from within the Docker instance
+cp $HOME/.awscredentials $HOME/.aws/credentials
+yarn
+yarn deploy
