+++ package.json
@@ -1,11 +1,16 @@
 {
   "name": "@code-dot-org/p5-replay",
   "version": "0.0.0",
-  "description": "Node.js .gif generator for p5.js sketches",
-  "main": "index.js",
+  "description": "Node.js .gif generator for p5.js sketches; AWS Lambda/Serverless setup based on https://github.com/kvaggelakos/serverless-ffmpeg",
   "scripts": {
     "postinstall": "cd node_modules; cp @code-dot-org/p5.play/lib/p5.play.js p5/lib/addons",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "deploy": "serverless deploy",
+    "stack-create": "sls deploy -v",
+    "stack-destroy": "sls remove -v",
+    "logs": "sls logs -f ffmpeg -t",
+    "test": "ava test/**/*.test.js",
+    "test-watch": "npm run test -- --watch",
+    "test-ci": "npm run test -- --verbose"
   },
   "repository": {
     "type": "git",
@@ -16,7 +21,12 @@
   "dependencies": {
     "@code-dot-org/p5.play": "1.2.2-cdo",
     "canvas": "^1.6.11",
-    "ffmpeg-static": "2.3.0",
-    "p5": "^0.7.1"
+    "p5": "^0.7.1",
+    "aws-sdk": "^2.173.0"
+  },
+  "devDependencies": {
+    "ava": "^0.24.0",
+    "fs-extra": "^5.0.0",
+    "serverless-s3-remover": "^0.4.1"
   }
 }
